requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-ckeditor==6.7.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-filter==24.2
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-modeltranslation==0.19.5
django-redis==5.4.0
django-rosetta==0.10.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
polib==1.2.0
psycopg==3.2.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
setuptools==70.3.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
